Map Screen

+++ app/src/main/java/com/aajogo/jogo/photosapp/data/models/data/ImageDto.kt
@@ -3,6 +3,6 @@ package com.aajogo.jogo.photosapp.data.models.data
 data class ImageDto(
     val base64Image: String,
     val date: Int,
-    val lat: Int,
-    val lng: Int
+    val lat: Double,
+    val lng: Double
 )
